Makes backends more pluggable, adds dynamodb backend
Previously, the `bucket_name` variable was overloaded depending on
the backend type. For S3, it was of course the actual bucket name.
But for SSM, it was used as the path to prefix the key. This was a
bit confusing and unexpected as a user. It also introduced an odd
question when trying to add backend: Do we add new/unique variables,
or continue overloading the (now) misnamed `bucket_name` variable?

This PR eliminates that overloaded variable by creating object-typed
variables for each type of backend. Now, each backend type may support
attributes distinct to that backend.

In addition, as a major release, the outputs are modified in a backwards-
incompatible manner. They were a list of objects, which made no sense
in retrospect. That was an artifact of inexperience with terraform.
This PR updates the outputs to be a map of key name to the keystore
object in the specified backend.

Finally, DynamoDB is added as a supported keystore backend. The DynamoDB
Table must already exist, and it must have a Partition Key of `Key`.
Values will be added to the attribute `Value`.